requirements.txt
fastapi==0.115.0
uvicorn==0.30.3

aiomysql==0.2.0
bcrypt==4.1.3
PyJWT==2.9.0
python-dotenv==1.0.1